# swagger.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: API documentation example
  version: 1.0.0
paths:
  /items:
    get:
      summary: Get all items
      responses:
        200:
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: 
                      type: integer
                    name:
                      type: string
              
    post:
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                name: 
                  type: string
                  example: Example Item
      responses:
        201:
          description: Created Item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  
  /items/{id}:
    put:
      summary: Update an item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Example Item
      responses:
        200:
          description: Updated item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string

    delete:
      summary: Delete an item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        204:
          description: No content